% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{add_stream_col}
\alias{add_stream_col}
\title{Add Stream Column}
\usage{
add_stream_col(data)
}
\arguments{
\item{data}{Data frame of Ascent data, generated from read_ascent_csv()}
}
\description{
This function adds a "Stream" column to an existing Ascent data frame.
This column is a factor with two levels, "Stream 1" and "Stream 2".
"Stream 1" is assigned if the batch name contains "_A_", and "Stream 2"
is assigned if the batch name contains "_B_".
}
\examples{
add_stream_col(data)
}
\keyword{ascent}
