% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_transitions.R
\name{asc_transitions}
\alias{asc_transitions}
\title{Identify Functional Entity Transitions and Driver Species}
\usage{
asc_transitions(x, top_n = 3)
}
\arguments{
\item{x}{An object of class \code{ascfcd} (paired) or \code{ascfcd_pw} (pairwise).}

\item{top_n}{Integer. Number of top contributing traits to consider (default: 3).}
}
\value{
A structured list of class \code{asc_trans}.
}
\description{
Decomposes the critical axes of change into specific trait shifts and identifies
the taxonomic drivers using Community Weighted Means (CWM) and Pearson correlation.
Automatically handles binary and continuous traits.
}
