% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCENT.R
\name{check_dimensions}
\alias{check_dimensions}
\title{Validate dimensions and feature names of an ASCENT object}
\usage{
check_dimensions(object)
}
\arguments{
\item{object}{An ASCENT S4 object.}
}
\value{
\code{TRUE} if valid, otherwise a character vector of error messages.
}
\description{
Checks that the RNA and ATAC matrices have the same number of cells
(columns) and that all genes and peaks referenced in \code{peak.info}
exist in the corresponding matrix rownames. Also warns if the ATAC
matrix has fewer rows than the RNA matrix (suggesting transposed input).
}
\details{
This function is used as the validity method for the ASCENT S4 class
and is called automatically on object creation.
}
