% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{clean_ascent_cols}
\alias{clean_ascent_cols}
\title{Clean Ascent Columns}
\usage{
clean_ascent_cols(data, posix = FALSE)
}
\arguments{
\item{data}{Data frame of Ascent data generated from read_ascent_csv()
with all columns present and column names unmodified.}

\item{posix}{Optionally converts AcqTime to POSIXct. Default is FALSE.}
}
\description{
This function takes a data frame generated from Ascent csv files and
selects only the relevant columns, and renames the columns so that they
are easier to access. Removes extraneous characters from the "Time" column.
Coerces various columns to factors.
}
\examples{
clean_ascent_cols()
}
\keyword{ascent}
