% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{clean_ascent_csv}
\alias{clean_ascent_csv}
\title{Clean Ascent CSV}
\usage{
clean_ascent_csv(batch_names, posix = TRUE)
}
\arguments{
\item{batch_names}{List of batch names}
}
\description{
Reads Ascent csv, generates a data frame that is easy to manipulate.
}
\examples{
clean_ascent_csv(batch_names, posix = TRUE)
}
\keyword{ascent}
