% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{evaluate_outliers}
\alias{evaluate_outliers}
\title{Identify Outliers}
\usage{
evaluate_outliers(data)
}
\arguments{
\item{data}{Data frame processed by clean_ascent_csv()}
}
\description{
Updates the "Include" column of an Ascent data frame with FALSE if the point is an outlier
Points are marked as outliers if they are less than Q1 - 1.5 * IQR or greater than Q3 + 1.5*IQR
}
\examples{
evaluate_outliers(data)
}
\keyword{ascent}
