% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ascfcd_pw}
\alias{plot.ascfcd_pw}
\title{Plot Pairwise Functional Spatial Divergence}
\usage{
\method{plot}{ascfcd_pw}(x, contrast, type = c("both", "pcoa", "leverage"), n_sp = 10, ...)
}
\arguments{
\item{x}{An object of class \code{ascfcd_pw}.}

\item{contrast}{Character. The contrast to plot (e.g., "SiteA_vs_SiteB").}

\item{type}{Character. What to plot? \code{"pcoa"}, \code{"leverage"}, or \code{"both"}.}

\item{n_sp}{Integer. Number of top species to display. Default is 10.}

\item{...}{Additional graphical arguments.}
}
\value{
A ggplot object.
}
\description{
Visualizes spatial functional restructuring (Centroid & Hull) between two communities.
}
