% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ascfcd_transitions}
\alias{plot.ascfcd_transitions}
\title{Plot Functional Leverage}
\usage{
\method{plot}{ascfcd_transitions}(x, contrast, n_sp = 10, ...)
}
\arguments{
\item{x}{An object of class \code{ascfcd_transitions}.}

\item{contrast}{Character. The contrast to plot.}

\item{n_sp}{Integer. Number of top species to display. Default is 10.}

\item{...}{Additional graphical arguments.}
}
\value{
A ggplot object.
}
\description{
Visualizes the top species driving functional centroid displacement.
}
