% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{plot_IS}
\alias{plot_IS}
\title{Plot Internal Standard}
\usage{
plot_IS(data)
}
\arguments{
\item{data}{Data frame of Ascent data, generated from read_ascent_csv(), cleaned with
clean_ascent_cols(posix = TRUE)}
}
\description{
Generates a plot of IS peak area vs acquisition time.
}
\examples{
plot_IS(data)
}
\keyword{ascent}
