% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{read_ascent_csv}
\alias{read_ascent_csv}
\title{Read Ascent CSVs}
\usage{
read_ascent_csv(batch_names)
}
\arguments{
\item{batch_names}{List of batch names.}
}
\description{
This function reads one or more Ascent csv files into a single data frame.
}
\examples{
read_ascent_csv()
}
