% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{summarize_cals}
\alias{summarize_cals}
\title{Create Summary Table of Calibrator Verification Data}
\usage{
summarize_cals(data, exclude_outliers = TRUE)
}
\arguments{
\item{data}{Data frame processed by clean_ascent_csv() and evaluate_outliers()}

\item{exclude_outliers}{defaults to TRUE. Will not include data in summary if data$Outlier is TRUE (as determined by IQR)}
}
\description{
Creates a summary table including Mean, SD, %CV, %Dev from Target, and n for each calibrator level for each analyte.
}
\examples{
summarize_cals(data)
}
\keyword{ascent}
