% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.asc_fe}
\alias{summary.asc_fe}
\title{Print summary for asc_fe objects}
\usage{
\method{summary}{asc_fe}(object, ...)
}
\arguments{
\item{object}{An object of class asc_fe}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary for asc_fe objects
}
