% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.asc_trans}
\alias{summary.asc_trans}
\title{Print summary for asc_trans objects}
\usage{
\method{summary}{asc_trans}(object, ...)
}
\arguments{
\item{object}{An object of class asc_trans}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary for asc_trans objects
}
