% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd_entities}
\alias{summary.ascfcd_entities}
\title{Summary for Functional Entities Classification}
\usage{
\method{summary}{ascfcd_entities}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ascfcd_entities}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the classification data frame.
}
\description{
Provides an overview of the species clustering into functional entities.
}
