% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd_pw}
\alias{summary.ascfcd_pw}
\title{Summary for Pairwise Functional Spatial Divergence}
\usage{
\method{summary}{ascfcd_pw}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ascfcd_pw}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the pairwise results data frame.
}
\description{
Provides a comprehensive overview of the spatial network ASC-CFD analysis.
}
