% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd_transitions}
\alias{summary.ascfcd_transitions}
\title{Summary for Functional Leverage Analysis}
\usage{
\method{summary}{ascfcd_transitions}(object, n_sp = 5, ...)
}
\arguments{
\item{object}{An object of class \code{ascfcd_transitions}.}

\item{n_sp}{Integer. Number of top species to display per contrast. Default is 5.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the object.
}
\description{
Provides an overview of the species leverage for each contrast.
}
