% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascent.R
\name{add_inst_col}
\alias{add_inst_col}
\title{Add Instrument Column}
\usage{
add_inst_col(data)
}
\arguments{
\item{data}{Data frame of Ascent data, generated from read_ascent_csv()}
}
\description{
Adds an "Instrument" column to a data frame resulting from read_ascent_csv().
}
\details{
Extracts characters after final underscore in batch name
}
\examples{
add_inst_col(data)
}
\keyword{ascent}
