% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_pairwise.R
\name{asc_pairwise}
\alias{asc_pairwise}
\title{Pairwise Functional Spatial Divergence}
\usage{
asc_pairwise(
  traits,
  abund,
  dist_method = "gower",
  dim_retention = c("variance", "broken_stick"),
  var_tol = 0.8,
  na.rm = TRUE
)
}
\arguments{
\item{traits}{A data frame or matrix of functional traits.}

\item{abund}{A data frame or matrix of species abundances.}

\item{dist_method}{Distance metric for the trait matrix. Default is "gower".}

\item{dim_retention}{Character. Method for dimensionality reduction.}

\item{var_tol}{Numeric. Proportion of cumulative variance to retain. Default is 0.80.}

\item{na.rm}{Logical. Remove missing values? Default is TRUE.}
}
\value{
An S3 object of class \code{ascfcd_pw}.
}
\description{
Calculates spatial functional divergence across three layers: position (Delta C), dispersion (FDis), and volume (FRic).
}
