% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_pairwise.R
\name{asc_pairwise}
\alias{asc_pairwise}
\title{Compute Pairwise ASC-FCD Network}
\usage{
asc_pairwise(traits, abund, dist_method = "jaccard")
}
\arguments{
\item{traits}{A data frame or matrix of species traits (rows = species, columns = traits).}

\item{abund}{A data frame or matrix of species abundances (rows = communities, columns = species).}

\item{dist_method}{Distance method for traits (default: "jaccard").}
}
\value{
An object of class \code{ascfcd_pw}.
}
\description{
Computes the functional displacement for all possible pairwise combinations
of communities (rows) in the abundance matrix.
}
