% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asc_transitions.R
\name{asc_transitions}
\alias{asc_transitions}
\title{Extract Functional Transitions and Species Leverage}
\usage{
asc_transitions(x, ...)
}
\arguments{
\item{x}{An object of class \code{ascfcd} or \code{ascfcd_pw}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of data frames containing species leverage for each contrast.
}
\description{
Calculates the multidimensional Functional Leverage of each species,
quantifying their direct contribution to the positional shift (Layer 1 - Delta C)
of the ecosystem.
}
