% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ascfcd}
\alias{plot.ascfcd}
\title{Plot Multidimensional Functional Restructuring}
\usage{
\method{plot}{ascfcd}(x, contrast, type = c("both", "pcoa", "leverage"), n_sp = 10, ...)
}
\arguments{
\item{x}{An object of class \code{ascfcd}.}

\item{contrast}{Character. The exact name of the site/contrast to plot.}

\item{type}{Character. What to plot? \code{"pcoa"} (Trajectory & Hull), \code{"leverage"} (Drivers), or \code{"both"}. Default is \code{"both"}.}

\item{n_sp}{Integer. Number of top species to display in the leverage plot. Default is 10.}

\item{...}{Additional graphical arguments.}
}
\value{
A ggplot object (or patchwork object if \code{type = "both"}).
}
\description{
Visualizes the functional trajectory (Centroid shift) and volume shift (Convex Hull)
of a specific site/contrast in the PCoA space, alongside the species leverage.
}
