% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.ascfcd_entities}
\alias{plot.ascfcd_entities}
\title{Plot Functional Entities Dendrogram}
\usage{
\method{plot}{ascfcd_entities}(x, ...)
}
\arguments{
\item{x}{An object of class \code{ascfcd_entities}.}

\item{...}{Additional graphical arguments passed to \code{plot.hclust}.}
}
\description{
Plots the hierarchical clustering dendrogram of the functional entities.
}
