% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd}
\alias{summary.ascfcd}
\title{Print summary for ascfcd paired objects}
\usage{
\method{summary}{ascfcd}(object, ...)
}
\arguments{
\item{object}{An object of class ascfcd}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary for ascfcd paired objects
}
