% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd_base}
\alias{summary.ascfcd_base}
\title{Summary for Baseline Functional Entities}
\usage{
\method{summary}{ascfcd_base}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ascfcd_ent}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisibly returns the entities results data frame.
}
\description{
Provides an overview of the baseline functional topology of the evaluated entities.
}
