% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.ascfcd_pw}
\alias{summary.ascfcd_pw}
\title{Print summary for ascfcd pairwise objects}
\usage{
\method{summary}{ascfcd_pw}(object, ...)
}
\arguments{
\item{object}{An object of class ascfcd_pw}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print summary for ascfcd pairwise objects
}
